refactor zns
