+++ zns/include/zns/bounded.h
